refactor and add janitor
